requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.1
python-dotenv==1.0.1
pydantic==2.9.2

# Voice
deepgram-sdk==3.7.7
elevenlabs==1.9.0
pvporcupine==3.0.2
pyaudio==0.2.14
sounddevice==0.5.0
numpy==1.26.4
torch==2.4.1
torchaudio==2.4.1

# LLM
anthropic==0.40.0
openai==1.51.0
google-generativeai==0.8.3
httpx==0.27.2

# Memory
chromadb==0.5.18
neo4j==5.25.0
redis==5.1.1
aiosqlite==0.20.0

# Tools
playwright==1.48.0
pyautogui==0.9.54
apscheduler==3.10.4
aiohttp==3.10.10
beautifulsoup4==4.12.3

# Vision
opencv-python==4.10.0.84
Pillow==10.4.0
mss==9.0.2

# Agents
langgraph==0.2.38
langchain-core==0.3.15
